Refactor: Move class Testcase to testcase.py
